requirements.txt
APScheduler==3.6.3
asgiref==3.4.1
astroid==2.9.3
backcall==0.2.0
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
decorator==5.1.0
Django==3.2.8
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==11.3.0
flake8==4.0.1
idna==3.3
ipdb==0.13.9
ipython==7.29.0
isort==5.10.1
jedi==0.18.0
lazy-object-proxy==1.7.1
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.10.0
pylint==2.12.2
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
wcwidth==0.2.5
wrapt==1.13.3
